Modify `+stats` commands to use custom paginators, allowing display of all players for that season/event.
