+++ cogs/utils/paginator.py
@@ -1,7 +1,14 @@
 import asyncio
 import discord
 import inspect
+import textwrap
+
+from datetime import datetime
+
 from discord.ext.commands import Paginator as CommandPaginator
+
+from cogs.utils.emoji_lookup import misc
+from cogs.utils.formatters import CLYTable, get_render_type, events_time
 
 
 class CannotPaginate(Exception):
@@ -323,9 +330,285 @@ class MessagePaginator(Pages):
         return '\n'.join(p)
 
 
+
+class TablePaginator(Pages):
+    def __init__(self, ctx, data, title=None, page_count=1, rows_per_table=20):
+        super().__init__(ctx, entries=[i for i in range(page_count)], per_page=1)
+        self.table = CLYTable()
+        self.data = [(i, v) for (i, v) in enumerate(data)]
+        self.entries = [None for _ in range(page_count)]
+        self.rows_per_table = rows_per_table
+        self.title = title
+        self.message = None
+        self.ctx = ctx
+        if getattr(ctx, 'config', None):
+            self.icon_url = ctx.config.icon_url
+            self.title = ctx.config.title
+        else:
+            self.icon_url = ctx.guild.icon_url
+
+    async def get_page(self, page):
+        entry = self.entries[page - 1]
+        if entry:
+            return entry
+
+        if not self.message:
+            self.message = await self.channel.send('Loading...')
+        else:
+            await self.message.edit(content='Loading...', embed=None)
+
+        entry = await self.prepare_entry(page)
+        self.entries[page - 1] = entry
+        return self.entries[page - 1]
+
+    async def prepare_entry(self, page):
+        self.table.clear_rows()
+        base = (page - 1) * self.rows_per_table
+        data = self.data[base:base + self.rows_per_table]
+        for n in data:
+            self.table.add_row(n)
+
+        render = get_render_type(self.ctx.config, self.table)
+        return render()
+
+    async def get_embed(self, entries, page, *, first=False):
+        if self.maximum_pages > 1:
+            if self.show_entry_count:
+                text = f'Page {page}/{self.maximum_pages} ({len(self.entries)} entries)'
+            else:
+                text = f'Page {page}/{self.maximum_pages}'
+
+            self.embed.set_footer(text=text)
+
+        self.embed.description = entries
+
+        self.embed.set_author(
+            name=textwrap.shorten(self.title, width=240, placeholder='...'),
+            icon_url=self.icon_url
+        )
+
+        return self.embed
+
+    async def show_page(self, page, *, first=False):
+        self.current_page = page
+        entries = await self.get_page(page)
+        embed = await self.get_embed(entries, page, first=first)
+
+        if not self.paginating:
+            print('not paginating')
+            return await self.message.edit(content=None, embed=embed)
+
+        await self.message.edit(content=None, embed=embed)
+
+        if not first:
+            return
+
+        for (reaction, _) in self.reaction_emojis:
+            if self.maximum_pages == 2 and reaction in ('\u23ed', '\u23ee'):
+                # no |<< or >>| buttons if we only have two pages
+                # we can't forbid it if someone ends up using it but remove
+                # it from the default set
+                continue
+
+            await self.message.add_reaction(reaction)
+
 class SeasonStatsPaginator(Pages):
     def __init__(self, ctx, entries):
         super().__init__(ctx, entries=entries, per_page=1)
 
     async def get_embed(self, entries, page, *, first=False):
         return self.entries[page - 1]
+
+
+class BoardPaginator(TablePaginator):
+    def __init__(self, ctx, data, title, page_count=1, rows_per_table=20):
+        super().__init__(ctx, data, title=title, page_count=page_count,
+                         rows_per_table=rows_per_table)
+
+    def create_row(self, player, data):
+        player_data = data[player.tag]
+
+        if self.ctx.config.type == 'donation':
+            if self.ctx.config.render == 1:
+                row = [player_data[0], player_data[1]['donations'], player_data[1]['received'], player.name]
+            else:
+                row = [player_data[0], player_data[1]['donations'], player.name]
+        else:
+            if self.ctx.config.render == 1:
+                row = [player_data[0], player_data[1]['trophies'], player_data[1][2], player.name]
+            else:
+                row = [player_data[0], player_data[1]['trophies'], player.name]
+
+        self.table.add_row(row)
+
+    async def prepare_entry(self, page):
+        self.table.clear_rows()
+        base = (page - 1) * self.rows_per_table
+        data = self.data[base:base + self.rows_per_table]
+        data_by_tag = {n[1]['player_tag']: n for n in data}
+
+        tags = [n[1]['player_tag'] for n in data]
+        async for player in self.bot.coc.get_players(tags):
+            self.create_row(player, data_by_tag)
+
+        render = get_render_type(self.ctx.config, self.table)
+        return render()
+
+
+class TrophyPaginator(TablePaginator):
+    def __init__(self, ctx, data, title, page_count=1, rows_per_table=20):
+        super().__init__(ctx, data, title=title, page_count=page_count,
+                         rows_per_table=rows_per_table)
+
+    def create_row(self, player, data):
+        player_data = data[player.tag]
+        row = [player_data[0], player_data[1][1], player_data[1][2], player.name]
+        self.table.add_row(row)
+
+    async def prepare_entry(self, page):
+        self.table.clear_rows()
+        base = (page - 1) * self.rows_per_table
+        data = self.data[base:base + self.rows_per_table]
+        data_by_tag = {n[1]['player_tag']: n for n in data}
+
+        tags = [n[1]['player_tag'] for n in data]
+        async for player in self.bot.coc.get_players(tags):
+            self.create_row(player, data_by_tag)
+
+        render = get_render_type(self.ctx.config, self.table)
+        return render()
+
+
+class LogsPaginator(TablePaginator):
+    def __init__(self, ctx, data, title, page_count=1, rows_per_table=20):
+        super().__init__(ctx, data, title=title, page_count=page_count,
+                         rows_per_table=rows_per_table)
+
+    async def prepare_entry(self, page):
+        self.table.clear_rows()
+        base = (page - 1) * self.rows_per_table
+        data = self.data[base:base + self.rows_per_table]
+        for player_data in data:
+            if self.ctx.config.type == 'donation':
+                player_data = player_data[1]
+                time = events_time((datetime.utcnow() - player_data[3]).total_seconds())
+                row = [
+                    misc['donated'] if player_data[1] else misc['received'],
+                    player_data[1] if player_data[1] else player_data[2],
+                    player_data[4],
+                    time
+                    ]
+            else:
+                player_data = player_data[1]
+                time = events_time((datetime.utcnow() - player_data[3]).total_seconds())
+                row = [
+                    misc['trophygreen'] if player_data[1] > 0 else misc['trophyred'],
+                    player_data[1],
+                    player_data[4],
+                    time
+                ]
+            self.table.add_row(row)
+
+        if self.ctx.config.type == 'donation':
+            return f"{self.table.donation_log_command()}\nKey: {misc['donated']} - Donated," \
+                   f" {misc['received']} - Received"
+        else:
+            return f"{self.table.trophy_log_command()}\nKey: {misc['trophygreen']} - Trophies Gained," \
+                   f" {misc['trophyred']} - Trophies Lost"
+
+
+class StatsAttacksPaginator(TablePaginator):
+    def __init__(self, ctx, data, title, page_count=1, rows_per_table=20, key=''):
+        super().__init__(
+            ctx, data, title=title, page_count=page_count, rows_per_table=rows_per_table
+        )
+        self.key = key
+
+    def create_row(self, player, data):
+        player_data = data[player.tag]
+        self.table.add_row([player_data[0], player_data[1]['attacks'], player_data[1]['trophies'], player.name])
+
+    async def prepare_entry(self, page):
+        self.table.clear_rows()
+        base = (page - 1) * self.rows_per_table
+        data = self.data[base:base + self.rows_per_table]
+        data_by_tag = {n[1]['player_tag']: n for n in data}
+
+        tags = [n[1]['player_tag'] for n in data]
+        async for player in self.bot.coc.get_players(tags):
+            self.create_row(player, data_by_tag)
+
+        return self.table.trophyboard_attacks() + self.key
+
+
+class StatsDefensesPaginator(TablePaginator):
+    def __init__(self, ctx, data, title, page_count=1, rows_per_table=20, key=''):
+        super().__init__(
+            ctx, data, title=title, page_count=page_count, rows_per_table=rows_per_table
+        )
+        self.key = key
+
+    def create_row(self, player, data):
+        player_data = data[player.tag]
+        self.table.add_row([player_data[0], player_data[1]['defenses'], player_data[1]['trophies'], player.name])
+
+    async def prepare_entry(self, page):
+        self.table.clear_rows()
+        base = (page - 1) * self.rows_per_table
+        data = self.data[base:base + self.rows_per_table]
+        data_by_tag = {n[1]['player_tag']: n for n in data}
+
+        tags = [n[1]['player_tag'] for n in data]
+        async for player in self.bot.coc.get_players(tags):
+            self.create_row(player, data_by_tag)
+
+        return self.table.trophyboard_defenses() + self.key
+
+
+class StatsGainsPaginator(TablePaginator):
+    def __init__(self, ctx, data, title, page_count=1, rows_per_table=20, key=''):
+        super().__init__(
+            ctx, data, title=title, page_count=page_count, rows_per_table=rows_per_table
+        )
+        self.key = key
+
+    def create_row(self, player, data):
+        player_data = data[player.tag]
+        self.table.add_row([player_data[0], player_data[1]['gains'], player_data[1]['trophies'], player.name])
+
+    async def prepare_entry(self, page):
+        self.table.clear_rows()
+        base = (page - 1) * self.rows_per_table
+        data = self.data[base:base + self.rows_per_table]
+        data_by_tag = {n[1]['player_tag']: n for n in data}
+
+        tags = [n[1]['player_tag'] for n in data]
+        async for player in self.bot.coc.get_players(tags):
+            self.create_row(player, data_by_tag)
+
+        return self.table.trophyboard_attacks() + self.key
+
+
+class StatsDonorsPaginator(TablePaginator):
+    def __init__(self, ctx, data, title, page_count=1, rows_per_table=20, key=''):
+        super().__init__(
+            ctx, data, title=title, page_count=page_count, rows_per_table=rows_per_table
+        )
+        self.key = key
+
+    def create_row(self, player, data):
+        player_data = data[player.tag]
+        self.table.add_row([player_data[0], player_data[1]['donations'], player_data[1]['trophies'], player.name])
+
+    async def prepare_entry(self, page):
+        self.table.clear_rows()
+        base = (page - 1) * self.rows_per_table
+        data = self.data[base:base + self.rows_per_table]
+        data_by_tag = {n[1]['player_tag']: n for n in data}
+
+        tags = [n[1]['player_tag'] for n in data]
+        async for player in self.bot.coc.get_players(tags):
+            self.create_row(player, data_by_tag)
+
+        return self.table.donationboard_2() + self.key
+
